refactor: extract Copilot API resolver from cli.ts (#2624)
Extract resolveCopilotApiKey, deriveCopilotApiTargetFromProviderBaseUrl,
deriveCopilotApiBasePathFromProviderBaseUrl, and resolveCopilotApiRouting
into src/copilot-api-resolver.ts (84 lines).

cli.ts re-exports all symbols for backwards compatibility.

All 32 cli tests pass. Build passes.

Closes #2624
